实现了分类功能,且替换为可视化可实现
